move custom object definitions to a separate file
