+++ gentoo_update/__init__.py
@@ -1,6 +1,15 @@
+from ._version import __version__
 from .gentoo_update import add_prefixes, create_cli, main
 from .shell_runner import ShellRunner
 from .parser import Parser
 from .reporter import Reporter
 from .notifier import Notifier
-from ._version import __version__
+from .report_objects import (
+        PackageInfo,
+        UpdateSection,
+        PretendError,
+        PretendSection,
+        DiskUsageStats,
+        DiskUsage,
+        LogInfo,
+)
